refactor: consolidate duplicate VersionTags table definition
